Added new test for trying to manipulate text
Added a new test that checks what happens if you add a number
and some string.  I've decided that I expect those requests to
return an HTTP 400 error.  Currently, the library does not do that,
so this is a case where the test was written before the actual
functional code.