Added new test file for post-minus
This new file will test the functionality of a new method
that has not yet been written.  It will use the POST method
of HTTP and it will perform the subtraction operation on
integers.  Still haven't figure out all the details.

This will definitely fail. Not sure if it will compile.